feat(ingestion): provide externe function to reuse
